requirements.txt
mkdocs-material
google-cloud-computeazure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity